fix: use bigint for stripe current period end and add await for updating Stripe subscription

+++ migrations/0000_init_db.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS "organization" (
 	"stripe_subscription_id" text,
 	"stripe_subscription_price_id" text,
 	"stripe_subscription_status" text,
-	"stripe_subscription_current_period_end" integer,
+	"stripe_subscription_current_period_end" bigint,
 	"updated_at" timestamp DEFAULT now() NOT NULL,
 	"created_at" timestamp DEFAULT now() NOT NULL
 );
